Renaming backend->server, frontend->client
