+++ src/log.js
@@ -2,6 +2,6 @@ import minilog from 'minilog'
 
 minilog.enable();
 
-const log = typeof window !== 'undefined' ? minilog('client') : minilog('backend');
+const log = typeof window !== 'undefined' ? minilog('client') : minilog('server');
 
 export default log;
